Fix dangling reference in itinerary

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -24,6 +24,9 @@ public class DeleteCommand extends Command {
             + "Example: " + COMMAND_WORD + " 1";
 
     public static final String MESSAGE_DELETE_ATTRACTION_SUCCESS = "Deleted Attraction: %1$s";
+    public static final String MESSAGE_ATTRACTION_IN_ITINERARY =
+            "Cannot delete attraction as it is referenced in one or more itineraries. "
+            + "Please remove it from all itineraries first.";
 
     private final Index targetIndex;
 
@@ -41,6 +44,11 @@ public class DeleteCommand extends Command {
         }
 
         Attraction attractionToDelete = lastShownList.get(targetIndex.getZeroBased());
+
+        if (model.isAttractionInAnyItinerary(attractionToDelete)) {
+            throw new CommandException(MESSAGE_ATTRACTION_IN_ITINERARY);
+        }
+
         model.deleteAttraction(attractionToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_ATTRACTION_SUCCESS, Messages.format(attractionToDelete)));
     }
